Refactor Storage to allow different backends
I am unable to run Redis in my environment, but am able to run Couchbase
and also use local disk storage.

I refactored the StorageMixin so that it's abstract and a new setting
named STORAGE_BACKEND defines which backend actually handles the
persisting of the values.

This doesn't introduce anything that's not backwards compatible and
allows existing deployments of Will to run unchanged.

As part of this I refactored the requirements so that there could be a
requirements file for couchbase as well as redis.