Fix a bug when loading rooms using the Hipchat V2 API
Currently, when retrieving rooms via the V2 API, if the max_results are
returned, the start-index is set to max_results and another request is
made to retrieve the next 1000 results. However, if another full page of
max_results are returned... an infinite loop occurs.

The fix is to add max_results to the start index parameter each time the
number of returned results is equal to max_results.